ufc_prediction_app/requirements.txt
# UFC Prediction App - Production Dependencies
# =============================================

# === Core Framework ===
streamlit>=1.29.0

# === Data Processing ===
pandas>=2.0.0
numpy>=1.24.0

# === Machine Learning ===
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.43.0
joblib>=1.3.0

# === MLOps / Experiment Tracking ===
mlflow>=2.9.0

# === Web Scraping ===
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# === Visualization ===
plotly>=5.18.0
altair>=5.0.0

# === LLM Integration ===
groq>=0.4.0

# === Utilities ===
python-dotenv>=1.0.0
tenacity>=8.2.0
pyyaml>=6.0.0

# === Optional (Kaggle data download) ===
kaggle>=1.5.0
